requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
build==1.2.2.post1
    # via pip-tools
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via
    #   sounddevice
    #   soundfile
click==8.1.8
    # via pip-tools
colorlog==6.9.0
    # via -r requirements.in
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.8.2
    # via openai
numpy==2.2.1
    # via
    #   -r requirements.in
    #   soundfile
openai==1.59.6
    # via -r requirements.in
opencc-python-reimplemented==0.1.7
    # via -r requirements.in
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.10.5
    # via openai
pydantic-core==2.27.2
    # via pydantic
pynput==1.7.7
    # via -r requirements.in
pyobjc-core==10.3.2
    # via
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-cocoa
    #   pyobjc-framework-coretext
    #   pyobjc-framework-quartz
pyobjc-framework-applicationservices==10.3.2
    # via pynput
pyobjc-framework-cocoa==10.3.2
    # via
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-coretext
    #   pyobjc-framework-quartz
pyobjc-framework-coretext==10.3.2
    # via pyobjc-framework-applicationservices
pyobjc-framework-quartz==10.3.2
    # via
    #   pynput
    #   pyobjc-framework-applicationservices
    #   pyobjc-framework-coretext
pyperclip==1.9.0
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyqt5==5.15.11
    # via -r requirements.in
pyqt5-qt5==5.15.16
    # via pyqt5
pyqt5-sip==12.16.1
    # via pyqt5
python-dotenv==1.0.0
    # via -r requirements.in
six==1.17.0
    # via pynput
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sounddevice==0.5.1
    # via -r requirements.in
soundfile==0.13.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
